requirements.txt
requests==2.31.0
openai==1.21.0
python-dotenv==1.0.1
streamlit==1.33.0
pydantic==2.6.4 